+++ src/components/Inputs/Checkbox/Checkbox.vue
@@ -66,7 +66,7 @@
   </div>
 </template>
 <script>
-import { inputMixins } from 'utils/inputMixins';
+import { inputMixins } from '@/utils/inputMixins';
 /**
  * Displays one or more checkboxes
  * @niceName Checkboxes
@@ -190,7 +190,3 @@ export default {
   },
 };
 </script>
-
-<style lang="scss">
-  @import '../../../assets/styles/scss/inputs.scss';
-</style>