+++ src/phila-ui.js
@@ -9,9 +9,12 @@ import Dropdown from './components/Inputs/Dropdown';
 import LangSelector from './components/LangSelector';
 import Textbox from './components/Inputs/Textbox';
 import TextArea from './components/Inputs/TextArea';
+import Phone from './components/Inputs/Phone';
 import TabsNav from './components/TabsNav';
 import MobileNav from './components/MobileNav';
 import DropdownNav from './components/DropdownNav';
+import Tooltip from './components/Tooltip';
+import Modal from './components/Modal';
 import ResponsiveHelpers from './plugins/ResponsiveHelpers';
 
 /**
@@ -48,10 +51,13 @@ export {
   Dropdown,
   Textbox,
   TextArea,
+  Phone,
   TabsNav,
   MobileNav,
   DropdownNav,
+  Tooltip,
+  Modal,
   ResponsiveHelpers,
-  install, // Install Helper for Vue
+  install,
 };